{"schema":"decodex/commit/1","summary":"Modularize capture host support","authority":"manual"}
