+++ native/macos-host/Sources/RsnapNativeHostKit/LiveChromeRefreshTelemetryKey.swift
@@ -1,0 +1,7 @@
+struct LiveChromeRefreshTelemetryKey: Equatable {
+	let targetHz: Int
+	let hudGlassEnabled: Bool
+	let hudGlassMode: String
+	let liquidGlassStyle: String
+	let liquidGlassAvailable: Bool
+}
