+++ native/macos-host/Sources/RsnapNativeHostKit/NativeHostTextMetrics.swift
@@ -1,0 +1,7 @@
+import AppKit
+
+extension String {
+	func size(using font: NSFont) -> CGSize {
+		(self as NSString).size(withAttributes: [.font: font])
+	}
+}
